+++ internal/pkg/cep_response.go
@@ -1,4 +1,4 @@
-package handler
+package pkg
 
 import (
 	"net/http"
