+++ internal/pkg/cep_search_test.go
@@ -1,4 +1,4 @@
-package handler
+package pkg
 
 import (
 	"testing"
