+++ internal/pkg/handler.go
@@ -1,4 +1,4 @@
-package handler
+package pkg
 
 import (
 	"encoding/json"
